+++ app/src/main/java/com/plaspa/weatherapp/model/Wind.kt
@@ -3,4 +3,4 @@ package com.plaspa.weatherapp.model
 /**
  * Created by Pedro on 13/03/2019.
  */
-class Wind(var speed: Double, var deg: Int, var gust: Double)
+class Wind(var speed: Double, var deg: Double, var gust: Double)
